supabase: use intermediate docker container to run supabase with docker compose
Replace swag with a normal nginx, because ssl for local instances is tricky.
Split the db files into init-scripts and migrations as this is the way the supabase/postgres
reads them.
Add numbers to the init scripts and migrations, that they are run last.
Use volumes instead of bind mounts, because localstack does not fully support docker in docker,
and we don't want to mess up the host file system.

Issue: #23