cd: add support for macos running arm processors
We did not get buildkit to work on macOS, so we cannot use the --link feature.

Also add g++, pyhton3 and make to the deps part.
Either node-gyp does not ship the needed arm binaries,
or the node:18-alpine image does not include g++, python3 and make in the arm image, but in the amd image.