mail-function: configure eslint more or less the same as in app
Use extends "plugin:prettier/recommended" to apply the prettier eslint rules
that the eslint rules don't conflict with prettier.
(As described in the docs: https://www.npmjs.com/package/eslint-plugin-prettier/v/4.2.1)
We don't have yet different rules for different folders, so we can put
it at the top level.

ALso verify that lint was applied in the ci.