mail-function: refactor to pure functions
In place updates of the parameters is not needed, and increases the cognitive
load on the developer.
Also use templated strings instead of string concatenation.
Use for of loop instead of self made for int i bla bla.
Convert unnecessary let to const.
Add error handling to loadDBData.
Expand abbreviation of parameters.