app: add authentication, sign up and profile capabilities
For now just add the pages, we will add ACL later.
Followed the guide of: https://supabase.com/docs/guides/auth/auth-helpers/nextjs

This was a little more involved because the order of the
migrations applied is different in the supabase development setup
and production setups (in cd.yml and development-setup.yml).
In the development setup the gotrue migrations happen first, and then
our migrations and the seed.sql is executed.
In the production setup (in cd.yml and development-setup.yml)
our migrations and seed.sql comes first, and then the gotrue migrations.

Issue: #11